third_party/blink/web_tests/external/wpt/svg/embedded/attr-image-fetchpriority-expected.txt

This is a testharness.js-based test.
[FAIL] fetchpriority attribute on <image> elements should reflect valid IDL values
  assert_equals: high fetchPriority is a valid IDL value on the image element expected (string) "high" but got (undefined) undefined
[FAIL] default fetchpriority attribute on <image> elements should be 'auto'
  assert_equals: expected (string) "auto" but got (undefined) undefined
Harness: the test ran to completion.
